pyproject.toml
[tool.poetry]
name = "locust-api-load-tester"
version = "0.1.0"
description = "Load-stress testing for API"
authors = ["Andrey K <andrey0kozyrev0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
locust = "^2.32.9"
ruff = "^0.9.6"
mimesis = "^18.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
